refactor: add ProposalVoteButton
